Don't sort lower and upper case differently
Tandoor and its recipes easily results in mixed casing of items on the shopping list, for example if a recipe calls for "milk" and not "Milk" its possible that all shopping list entries for milk will be either case. When displaying on a shopping list it seems unlikely anyone would want "milk" to be displayed after "Salad" while "Milk" would be before "Salad".